Fix min_inverse_rank_col_pref_large when called with masked value.
Calling the script with a mask value can cause each column to have a
different number of masked values. Since the
min_inverse_rank_col_pref_large function must know the position of the
last non-masked value, to apply a weight of 1 to it, the function has to
calculate the weights per column, depending on the specific number of
non-masked values in each column.

Previously, the function used to largest number of rows from all
columns, resulting in smaller column sums for columns with less rows,
since the factor 1 was applied to a masked value.